+++ src/main/java/com/gotcha/earlytable/domain/waiting/dto/WaitingSimpleResponseDto.java
@@ -1,18 +1,13 @@
 package com.gotcha.earlytable.domain.waiting.dto;
 
-import com.gotcha.earlytable.domain.party.entity.PartyPeople;
 import com.gotcha.earlytable.domain.waiting.entity.Waiting;
-import com.gotcha.earlytable.global.enums.WaitingStatus;
-import com.gotcha.earlytable.global.enums.WaitingType;
 import lombok.Getter;
 
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 public class WaitingSimpleResponseDto {
     Long waitingId;
-    Integer waitingNumber;
+    Long waitingNumber;
     Integer personnelCount;
     String phoneNumber;
